DP-6517: Demonstrate alert providing transform
