+++ styleguide/source/assets/data/emergency-alerts-override.json
@@ -1,0 +1,38 @@
+{
+  "emergencyAlerts": {
+    "id":"GUID198767819323444",
+    "buttonAlert": {
+      "hideText": "Hide",
+      "showText": "Show",
+      "text": "Alerts"
+    },
+    "emergencyHeader": {
+      "title": "The State is experiencing severe weather due to the winter storm Paula."
+    },
+    "alertsOverride": [{
+      "message": "MassPort is operating with limited flights due to road closures.",
+      "timeStamp": "02.15.16, 5:00 p.m.",
+      "link": {
+        "href": "#",
+        "text":"Read full alert",
+        "chevron": "true"
+      }
+    },{
+      "message": "The MTBA is operating at a lighter service level, and experiencing delays.",
+      "timeStamp": "02.15.16, 3:00 p.m.",
+      "link": {
+        "href": "#",
+        "text":"Read full alert",
+        "chevron": "true"
+      }
+    },{
+      "message": "The Governor directs non-emergency executive branch state employees to not come in to work.",
+      "timeStamp": "02.13.16, 10:00 a.m.",
+      "link": {
+        "href": "#",
+        "text":"Read full alert",
+        "chevron": "true"
+      }
+    }]
+  }
+}
