+++ styleguide/source/assets/js/index.js
@@ -1,6 +1,8 @@
 import                            "./helpers/jQueryExtend.js";
+import                            "./helpers/setAlertsJQueryExtend.js";
 import                            "./helpers/handlebarsHelpers.js";
 import ajaxAlerts                 from "./modules/ajaxAlerts.js";
+import ajaxAlertsOverride         from "./modules/ajaxAlertsOverride.js";
 import accordions                 from "./modules/accordions.js";
 import googleMap                  from "./modules/googleMap.js";
 import back2top                   from "./modules/back2top.js";
